refactor(metrics-exporter): introduce the otlp-proto-exporter-base package.
