+++ .gitmodules
@@ -1,9 +1,6 @@
-[submodule "experimental/packages/exporter-trace-otlp-proto/protos"]
-	path = experimental/packages/exporter-trace-otlp-proto/protos
-	url = https://github.com/open-telemetry/opentelemetry-proto.git
-[submodule "experimental/packages/opentelemetry-exporter-metrics-otlp-proto/protos"]
-	path = experimental/packages/opentelemetry-exporter-metrics-otlp-proto/protos
-	url = https://github.com/open-telemetry/opentelemetry-proto.git
 [submodule "experimental/packages/otlp-grpc-exporter-base/protos"]
 	path = experimental/packages/otlp-grpc-exporter-base/protos
+	url = https://github.com/open-telemetry/opentelemetry-proto.git
+[submodule "experimental/packages/otlp-proto-exporter-base/protos"]
+	path = experimental/packages/otlp-proto-exporter-base/protos
 	url = https://github.com/open-telemetry/opentelemetry-proto.git
