+++ experimental/packages/exporter-trace-otlp-proto/package.json
@@ -73,6 +73,7 @@
     "@opentelemetry/resources": "1.1.1",
     "@opentelemetry/sdk-trace-base": "1.1.1",
     "@opentelemetry/otlp-exporter-base": "0.27.0",
+    "@opentelemetry/otlp-proto-exporter-base": "0.27.0",
     "protobufjs": "^6.9.0"
   }
 }
