+++ experimental/packages/exporter-trace-otlp-proto/src/OTLPTraceExporter.ts
@@ -15,14 +15,13 @@
  */
 
 import { ReadableSpan, SpanExporter } from '@opentelemetry/sdk-trace-base';
-import { OTLPExporterNodeBase } from './OTLPExporterNodeBase';
 import {
   otlpTypes,
   toOTLPExportTraceServiceRequest
 } from '@opentelemetry/exporter-trace-otlp-http';
-import { ServiceClientType } from './types';
 import { getEnv, baggageUtils } from '@opentelemetry/core';
 import { appendResourcePathToUrlIfNotPresent, OTLPExporterNodeConfigBase } from '@opentelemetry/otlp-exporter-base';
+import { OTLPProtoExporterNodeBase, ServiceClientType } from '@opentelemetry/otlp-proto-exporter-base';
 
 const DEFAULT_COLLECTOR_RESOURCE_PATH = '/v1/traces';
 const DEFAULT_COLLECTOR_URL=`http://localhost:4318${DEFAULT_COLLECTOR_RESOURCE_PATH}`;
@@ -31,7 +30,7 @@ const DEFAULT_COLLECTOR_URL=`http://localhost:4318${DEFAULT_COLLECTOR_RESOURCE_P
  * Collector Trace Exporter for Node with protobuf
  */
 export class OTLPTraceExporter
-  extends OTLPExporterNodeBase<
+  extends OTLPProtoExporterNodeBase<
     ReadableSpan,
     otlpTypes.opentelemetryProto.collector.trace.v1.ExportTraceServiceRequest
   >
