+++ experimental/packages/otlp-proto-exporter-base/.eslintignore
@@ -1,0 +1,1 @@
+build
