+++ experimental/packages/otlp-proto-exporter-base/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+    "env": {
+        "mocha": true,
+        "commonjs": true,
+        "node": true,
+    },
+    ...require('../../../eslint.config.js')
+}
