+++ experimental/packages/otlp-proto-exporter-base/src/types.ts
