+++ experimental/tsconfig.json
@@ -54,6 +54,9 @@
       "path": "packages/otlp-grpc-exporter-base"
     },
     {
+      "path": "packages/otlp-proto-exporter-base"
+    },
+    {
       "path": "packages/otlp-transformer"
     },
     {
